Write Generic Text importer script
Switch to generic text importer in reception scene
