+++ Assets/Shared Scripts/GenericTextImporter.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 53f4033524499514c9c3d6263ce10b02
-timeCreated: 1519750920
+guid: 80ed8505cfc52744699c69fcf86c0668
+timeCreated: 1521295808
 licenseType: Free
 MonoImporter:
   externalObjects: {}
